requirements.txt
Django==4.2.5
djangorestframework==3.14.0
mysql-connector-python
gunicorn
psycopg2
Pillow
drf-yasg  
whitenoise
django-webpack-loader
